footballAnalyst/requirement.txt
# requirements.txt

aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.17.0
asgiref==3.8.1
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-ratelimit==4.1.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
multidict==6.1.0
nltk==3.9.1
openpyxl==3.1.5
pillow==11.0.0
propcache==0.2.1
python-dotenv==1.0.1
ratelimit==2.2.1
regex==2024.11.6
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
textblob==0.18.0.post0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
